Add note on DBD APIs to Module Developer Changes in new_features. Update cast in mod_dbd.c as suggested by jorton.
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@291100 13f79535-47bb-0310-9956-ffa450edef68

+++ modules/database/mod_dbd.c
@@ -71,7 +71,7 @@ static const char *dbd_param(cmd_parms *cmd, void *cfg, const char *val)
     svr_cfg *svr = (svr_cfg*) ap_get_module_config
         (cmd->server->module_config, &dbd_module);
 
-    switch ((int) cmd->info) {
+    switch ((long) cmd->info) {
     case cmd_name:
         svr->name = val;
         /* loading the driver involves once-only dlloading that is
